chore: initial provider config
