+++ web/lib/providers/counter_provider.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class CounterProvider with ChangeNotifier {
+  int _count = 0;
+
+  int get count => _count;
+
+  void increment() {
+    _count++;
+    notifyListeners();
+  }
+}
